Use relative path for git submodules
This commit adds a relative path to all git submodules.
This makes sure that if a user copies over HTTPS the submodules
are also copied over HTTPS.
If the user copies over SSH the submodules are downloaded over SSH
This is useful in CI environments.

as per https://stackoverflow.com/questions/40841882/automatically-access-git-submodules-via-ssh-or-https

Note that this commit also has changed the README.md.
In my testing the pulp-nn-1d repository is not public, so
initializing the submodules with the --recursive flag results in
an error.
In my case this also meant that the public submodules were left in
a weird state. (submodules were initialized but all files were
deleted, so you would have to `git restore` each submodule).
Initializing every submodule separately should solve this issue.